Fix remaining signed/unsigned warnings * This seems to fix the 64 bit and signed/unsigned warnings in GWEN for   GCC and clang (Xcode). * size_t cast to 32 bit unsigned int. GWEN assumes 32 bit compilation.   This is good enough. GWEN would need a fair bit of work to make 64 compliant.

+++ gwen/include/Gwen/TextObject.h
@@ -94,9 +94,9 @@ namespace Gwen
             return m_Unicode;
         }
 
-        int length() const
+        unsigned int length() const
         {
-            return m_Unicode.length();
+            return (unsigned int)m_Unicode.length();
         }
 
     private:
